pavilion/forge/requirements.txt
anthropic>=0.40.0
slack_sdk>=3.27.0
google-api-python-client>=2.120.0
google-auth>=2.29.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=8.0.0
